avalanche-types: move types from avalanche-api
