+++ avalanche-ops-aws/Cargo.toml
@@ -12,7 +12,7 @@ avalanchego = { path = "../avalanchego" }
 aws = { path = "../aws" }
 aws-sdk-cloudformation = "0.9.0"
 aws-sdk-s3 = "0.9.0"
-clap = { version = "3.1.6", features = ["derive"] }
+clap = { version = "3.1.8", features = ["derive"] }
 coreth = { path = "../coreth" }
 crossterm = "0.23.1"
 dialoguer = "0.10.0"
