+++ avalanche-types/Cargo.toml
@@ -9,10 +9,12 @@ edition = "2021"
 bech32 = "0.8.1"
 bitcoin = "0.27.1"
 bytes = "1.1.0"
+chrono = "0.4.19"
 ethereum-types = "0.13.1"
 hex = "0.4.3"
 lazy_static = "1.4.0"
 log = "0.4.16"
+num-bigint = "0.4.3"
 openssl = "0.10.38"
 ripemd = "0.1.1"
 rust-embed = "6.3.0"
