+++ avalanche-types/README
@@ -1,0 +1,2 @@
+
+Avalanche protocol primitive types.
