+++ avalanche-types/artifacts/sample.genesis.json
