+++ avalanche-types/src/api/mod.rs
@@ -2,6 +2,5 @@ pub mod avm;
 pub mod eth;
 pub mod health;
 pub mod info;
-pub mod metrics;
-pub mod p;
-pub mod platform;
+pub mod jsonrpc;
+pub mod platformvm;
