+++ avalanche-types/src/metrics/mod.rs
@@ -1,0 +1,1 @@
+pub mod avalanchego;
