+++ avalanched-aws/Cargo.toml
@@ -12,7 +12,7 @@ avalanche-types = { path = "../avalanche-types" }
 avalanchego = { path = "../avalanchego" }
 aws = { path = "../aws" }
 aws-sdk-s3 = "0.9.0"
-clap = { version = "3.1.6", features = ["derive"] }
+clap = { version = "3.1.8", features = ["derive"] }
 env_logger = "0.9.0"
 log = "0.4.16"
 tempfile = "3.3.0"
