+++ coreth/src/lib.rs
@@ -1,2 +1,1 @@
 pub mod config;
-pub mod genesis;
